fix(OnBoardingScreen): fix errors
